move enum files for better structure
