+++ enum/science/composition.ts
