+++ enum/science/descriptive.ts
