+++ enum/science/generic.ts
