+++ enum/science/nuclear.ts
