+++ enum/science/reaction.ts
