+++ enum/system/condition.ts
