+++ enum/system/person.ts
