+++ enum/system/uncertainty.ts
