+++ model/collection/composition.ts
@@ -1,7 +1,7 @@
 import type { Brand, Expand } from 'devtypes/types/util';
-import type { CompositionType } from '../../enum/collection/composition';
-import type { ElementSymbol } from '../../enum/domain/element';
-import type { EntityType } from '../../enum/registry/system';
+import type { CompositionType } from '../../enum/science/composition';
+import type { ElementSymbol } from '../../enum/science/element';
+import type { DomainType } from '../../enum/system/domain';
 import type { Collection, Distinct } from '../base/modifier';
 import type { CompoundId } from '../domain/compound';
 import type { MineralId } from '../domain/mineral';
@@ -9,10 +9,10 @@ import type { MixtureId } from '../domain/mixture';
 
 type BaseComponent< T extends CompositionType > = Brand< {
   ref:
-    | { type: EntityType.ELEMENT, id: ElementSymbol }
-    | { type: EntityType.COMPOUND, id: CompoundId }
-    | { type: EntityType.MIXTURE, id: MixtureId }
-    | { type: EntityType.MINERAL, id: MineralId };
+    | { type: DomainType.ELEMENT, id: ElementSymbol }
+    | { type: DomainType.COMPOUND, id: CompoundId }
+    | { type: DomainType.MIXTURE, id: MixtureId }
+    | { type: DomainType.MINERAL, id: MineralId };
   charge?: number;
 }, T, 'type', true >;
 
@@ -31,11 +31,11 @@ type FractionComponent = Expand< BaseComponent< CompositionType.FRACTION > & {
   fraction: number;
 } >;
 
-export type CompositionCollection< E extends EntityType > = Collection< {
+export type CompositionCollection< E extends DomainType > = Collection< {
   components: Distinct<
-    E extends EntityType.COMPOUND ? StoichiometryComponent :
-    E extends EntityType.MINERAL ? StoichiometryComponent | RangeComponent :
-    E extends EntityType.MIXTURE ? FractionComponent :
+    E extends DomainType.COMPOUND ? StoichiometryComponent :
+    E extends DomainType.MINERAL ? StoichiometryComponent | RangeComponent :
+    E extends DomainType.MIXTURE ? FractionComponent :
     never
   >[];
 } >;
